another try

+++ grpc-interface-support/test/integration/test_integration_python.py
@@ -21,7 +21,8 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def clean_downloads():
-    os.removedirs(os.path.join(get_project_cache_dir(), "downloads"))
+    if os.path.exists(os.path.join(get_project_cache_dir(), "downloads")):
+        os.removedirs(os.path.join(get_project_cache_dir(), "downloads"))
 
 
 if not os.environ["VELOCITAS_TEST_LANGUAGE"] == "python":
