+++ sdk-installer/test/integration/test_integration_python.py
@@ -24,7 +24,8 @@ if not os.environ["VELOCITAS_TEST_LANGUAGE"] == "python":
 
 @pytest.fixture(autouse=True)
 def clean_downloads():
-    os.removedirs(os.path.join(get_project_cache_dir(), "downloads"))
+    if os.path.exists(os.path.join(get_project_cache_dir(), "downloads")):
+        os.removedirs(os.path.join(get_project_cache_dir(), "downloads"))
 
 
 def get_subdirs(path: str) -> List[str]:
